fix: consolidate files to satisfy pkg-file-count gate
