+++ docs/internal/processes/invocation-relevant-files.md
@@ -43,7 +43,7 @@ primary-result behavior.
   invocation input-source rules and the canonical pointers into packaged docs.
   `runInvocationModes` and `resolveRunFactoryPrompt` also treat `you run --named`
   as an invocation factory selector for positional/stdin text.
-- `pkg/config/builtin_tts_factory.go` owns the built-in `@you/tts` factory JSON
+- `pkg/config/layout.go` owns the built-in `@you/tts` factory JSON (`BuiltInTTSFactoryJSON`)
   registered from `builtInNamedFactoryCatalog` in `pkg/config/layout.go`.
 - `pkg/packagedfactories/tts/` owns packaged TTS invocation metadata shaping
   helpers used when MODEL_INVOKE work completes on the `execute-tts` workstation.
@@ -59,10 +59,9 @@ primary-result behavior.
 - `pkg/cli/run/packaged_tts_invocation.go` logs named-factory resolution context at
   the CLI boundary without recording packaged-factory metrics or logging submitted
   text or generated artifact bodies.
-- `pkg/service/model_catalog.go`, `pkg/service/model_catalog_invocation_wait.go`,
-  and `pkg/service/model_catalog_packaged_tts.go` own the session invocation wait
-  loop, packaged TTS loading/completion/failure logs, and packaged-factory metrics
-  while polling for primary results.
+- `pkg/service/model_catalog.go` owns the session invocation wait loop, packaged TTS
+  loading/completion/failure logs, and packaged-factory metrics while polling for
+  primary results.
 - `pkg/factory/subsystems/subsystem_transitioner.go` applies packaged TTS
   invocation metadata to terminal token `Content` for the `execute-tts` TTS
   MODEL_INVOKE workstation so primary-result selection returns JSON metadata
