+++ pkg/service/model_catalog.go
@@ -1,3 +1,5 @@
+// backendsizecheck:ignore-file model catalog owns packaged TTS invocation wait and metrics paths until dedicated service seams split.
+// pkgmaintcheck:ignore-file-lines model catalog owns packaged TTS invocation wait and metrics paths until dedicated service seams split.
 package service
 
 import (
@@ -878,4 +880,209 @@ func stringValue[T ~string](value *T) string {
 		return ""
 	}
 	return string(*value)
+}
+
+type invocationWaitTickResult struct {
+	done          bool
+	result        apisurface.FactoryInvocationResult
+	err           error
+	loggedLoading bool
+}
+
+func invocationWaitContext(ctx context.Context, timeoutMillis *int64) (context.Context, context.CancelFunc) {
+	if timeoutMillis != nil && *timeoutMillis > 0 {
+		return context.WithTimeout(ctx, time.Duration(*timeoutMillis)*time.Millisecond)
+	}
+	return ctx, func() {}
+}
+
+func (fs *FactoryService) isPackagedTTSSession(sessionID string) bool {
+	runtimeCfg, err := fs.sessionRuntimeConfig(sessionID)
+	return err == nil && runtimeCfg != nil && tts.IsPackagedFactory(runtimeCfg.FactoryConfig())
+}
+
+func (fs *FactoryService) processInvocationWaitTick(
+	waitCtx context.Context,
+	sessionID string,
+	input sessionInvocationWaitInput,
+	result apisurface.FactoryInvocationResult,
+	packagedTTSInvocation bool,
+	loggedPackagedLoading bool,
+) invocationWaitTickResult {
+	snapshot, err := fs.GetEngineStateSnapshotForSession(waitCtx, sessionID)
+	if err != nil {
+		waitResult, waitErr := fs.handleInvocationWaitError(result, err)
+		return invocationWaitTickResult{done: true, result: waitResult, err: waitErr}
+	}
+
+	worldState, err := fs.sessionInvocationWorldState(waitCtx, sessionID, snapshot.TickCount)
+	if err != nil {
+		waitResult, waitErr := fs.handleInvocationWaitError(result, err)
+		return invocationWaitTickResult{done: true, result: waitResult, err: waitErr}
+	}
+
+	activeWork := snapshotHasActiveWork(snapshot)
+	if packagedTTSInvocation && activeWork && !loggedPackagedLoading {
+		fs.logPackagedTTSInvocationLoading(sessionID, input)
+		loggedPackagedLoading = true
+	}
+
+	selection, selectionErr := invocations.ResolvePrimaryResult(invocations.PrimaryResultSelectionInput{
+		RequestID:        input.RequestID,
+		InvocationReturn: input.InvocationReturn,
+		WorldState:       worldState,
+	})
+	if selectionErr == nil {
+		if packagedTTSInvocation {
+			fs.logPackagedTTSInvocationCompleted(sessionID, input, selection)
+		}
+		return invocationWaitTickResult{
+			done:   true,
+			result: fs.handleInvocationSelectionSuccess(sessionID, input, selection),
+		}
+	}
+
+	primaryErr, ok := selectionErr.(*invocations.PrimaryResultError)
+	if !ok {
+		return invocationWaitTickResult{done: true, err: selectionErr}
+	}
+
+	if _, exists := worldState.WorkRequestsByID[input.RequestID]; exists && !activeWork {
+		return invocationWaitTickResult{
+			done:   true,
+			result: fs.resolveInvocationWaitTerminal(sessionID, input, worldState, packagedTTSInvocation, primaryErr),
+		}
+	}
+
+	return invocationWaitTickResult{loggedLoading: loggedPackagedLoading}
+}
+
+func (fs *FactoryService) resolveInvocationWaitTerminal(
+	sessionID string,
+	input sessionInvocationWaitInput,
+	worldState interfaces.FactoryWorldState,
+	packagedTTSInvocation bool,
+	primaryErr *invocations.PrimaryResultError,
+) apisurface.FactoryInvocationResult {
+	if packagedTTSInvocation {
+		if _, failure := tts.ClassifyInvocationWait(worldState, input.RequestID, false); failure != nil {
+			return fs.handlePackagedTTSInvocationFailure(sessionID, input, failure)
+		}
+	}
+	return fs.handleInvocationUnresolvedPrimary(sessionID, input, primaryErr)
+}
+
+func (fs *FactoryService) invocationWaitTimedOut(
+	sessionID string,
+	input sessionInvocationWaitInput,
+	result apisurface.FactoryInvocationResult,
+	waitErr error,
+) apisurface.FactoryInvocationResult {
+	terminalResult := invocationContextTerminalResult(result, waitErr)
+	fs.recordInvocationMetric(invocationMetricFailure, inputMetricLabels(input.InputSource))
+	fs.logInvocationTerminalResult(sessionID, input, terminalResult)
+	return terminalResult
+}
+
+func (fs *FactoryService) recordPackagedTTSInvocationMetric(
+	name string,
+	source invocations.InputSourceLabel,
+	extra map[string]string,
+) {
+	labels := mergeMetricLabels(
+		inputMetricLabels(source),
+		map[string]string{"packaged_factory": tts.PackagedFactoryName},
+		extra,
+	)
+	fs.recordInvocationMetric(name, labels)
+}
+
+func (fs *FactoryService) handlePackagedTTSInvocationFailure(
+	sessionID string,
+	input sessionInvocationWaitInput,
+	failure *tts.InvocationFailure,
+) apisurface.FactoryInvocationResult {
+	result := apisurface.FactoryInvocationResult{
+		RequestID: input.RequestID,
+		TraceID:   input.TraceID,
+		Status:    factoryapi.InvocationTerminalStatusFailed,
+		ErrorCode: failure.ErrorCode,
+		Message:   failure.Message,
+	}
+	fs.recordInvocationMetric(invocationMetricFailure, inputMetricLabels(input.InputSource))
+	fs.recordPackagedTTSInvocationMetric(tts.MetricPackagedFactoryFailure, input.InputSource, map[string]string{
+		"failure_class": failure.FailureClass,
+	})
+	if failure.FailureClass == tts.FailureClassModelNotReady {
+		fs.recordPackagedTTSInvocationMetric(tts.MetricPackagedFactoryNotReady, input.InputSource, nil)
+	}
+	fs.logger.Warn(
+		"packaged tts invocation failed",
+		packagedTTSInvocationLogFields(
+			sessionID,
+			input.InputSource,
+			input.InvocationReturn,
+			zap.String("request_id", input.RequestID),
+			zap.String("trace_id", input.TraceID),
+			zap.String("status", string(result.Status)),
+			zap.String("error_code", result.ErrorCode),
+			zap.String("failure_class", failure.FailureClass),
+			zap.String("readiness_outcome", failure.FailureClass),
+		)...,
+	)
+	return result
+}
+
+func (fs *FactoryService) logPackagedTTSInvocationLoading(
+	sessionID string,
+	input sessionInvocationWaitInput,
+) {
+	fs.logger.Info(
+		"packaged tts invocation loading",
+		packagedTTSInvocationLogFields(
+			sessionID,
+			input.InputSource,
+			input.InvocationReturn,
+			zap.String("request_id", input.RequestID),
+			zap.String("trace_id", input.TraceID),
+			zap.String("readiness_outcome", tts.FailureClassLoading),
+		)...,
+	)
+}
+
+func (fs *FactoryService) logPackagedTTSInvocationCompleted(
+	sessionID string,
+	input sessionInvocationWaitInput,
+	selection invocations.PrimaryResultSelection,
+) {
+	fs.recordPackagedTTSInvocationMetric(tts.MetricPackagedFactorySuccess, input.InputSource, map[string]string{
+		"readiness_outcome": tts.FailureClassSuccess,
+	})
+	fs.logger.Info(
+		"packaged tts invocation completed",
+		packagedTTSInvocationLogFields(
+			sessionID,
+			input.InputSource,
+			input.InvocationReturn,
+			zap.String("request_id", input.RequestID),
+			zap.String("trace_id", input.TraceID),
+			zap.String("status", string(factoryapi.InvocationTerminalStatusCompleted)),
+			zap.String("resolved_work_id", selection.WorkID),
+			zap.String("resolved_work_type", selection.WorkTypeName),
+			zap.String("readiness_outcome", tts.FailureClassSuccess),
+		)...,
+	)
+}
+
+func packagedTTSInvocationLogFields(
+	sessionID string,
+	source invocations.InputSourceLabel,
+	cfg *interfaces.InvocationReturnConfig,
+	extra ...zap.Field,
+) []zap.Field {
+	fields := invocationLogFields(sessionID, source, cfg,
+		zap.String("packaged_factory_name", tts.PackagedFactoryName),
+		zap.String("tts_backend", tts.BackendRuntimeLabel()),
+	)
+	return append(fields, extra...)
 }
